4.21

+++ osa4/controllers/blogs.js
@@ -1,3 +1,4 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
@@ -49,8 +50,20 @@ blogsRouter.post('/', async (request, response) => {
   //     })
   })
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id) 
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+          return response.status(401).json({ error: 'token invalid' })
+          }  
+    const user = await User.findById(decodedToken.id)
+    
+  if (blog.user.toString() === user.id.toString()) {
+    console.log("jeps")
+    await Blog.findByIdAndRemove(request.params.id) 
     response.status(204).end()
+  }
+
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
